group similar features in same file
