--- src/plugins/profiling/server/routes/search_topn_threads.ts
@@ -1,18 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0 and the Server Side Public License, v 1; you may not use this file except
- * in compliance with, at your election, the Elastic License 2.0 or the Server
- * Side Public License, v 1.
- */
-import type { DataRequestHandlerContext } from '../../../data/server';
-import type { IRouter } from '../../../../core/server';
-import { getRemoteRoutePaths } from '../../common';
-import { queryTopNCommon } from './search_topN';
-
-export function registerTraceEventsTopNThreadsSearchRoute(
-  router: IRouter<DataRequestHandlerContext>
-) {
-  const paths = getRemoteRoutePaths();
-  return queryTopNCommon(router, paths.TopNThreads, 'ThreadName');
-}
